+++ src/Tags/Model/TagsListOrderFields.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\SDK\Tags\Model;
+
+final class TagsListOrderFields
+{
+    public const TAG = 'tag';
+    public const SHORT_URLS_COUNT = 'shortUrlsCount';
+    public const VISITS_COUNT = 'visitsCount';
+
+    private function __construct()
+    {
+    }
+}