+++ test/ShlinkClientTest.php
@@ -22,6 +22,8 @@ use Shlinkio\Shlink\SDK\ShortUrls\Model\ShortUrlsFilter;
 use Shlinkio\Shlink\SDK\ShortUrls\Model\ShortUrlsList;
 use Shlinkio\Shlink\SDK\ShortUrls\ShortUrlsClientInterface;
 use Shlinkio\Shlink\SDK\Tags\Model\TagRenaming;
+use Shlinkio\Shlink\SDK\Tags\Model\TagsFilter;
+use Shlinkio\Shlink\SDK\Tags\Model\TagsWithStatsList;
 use Shlinkio\Shlink\SDK\Tags\TagsClientInterface;
 use Shlinkio\Shlink\SDK\Visits\Model\VisitsFilter;
 use Shlinkio\Shlink\SDK\Visits\Model\VisitsList;
@@ -79,7 +81,7 @@ class ShlinkClientTest extends TestCase
     public function listShortUrlsDelegatesCallToProperClient(): void
     {
         $listUrls = $this->shortUrlsClient->listShortUrls()->willReturn(
-            ShortUrlsList::forTupleLoader(fn () => [[], []]),
+            ShortUrlsList::forTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listShortUrls();
@@ -92,7 +94,7 @@ class ShlinkClientTest extends TestCase
     {
         $filter = ShortUrlsFilter::create();
         $listUrls = $this->shortUrlsClient->listShortUrlsWithFilter($filter)->willReturn(
-            ShortUrlsList::forTupleLoader(fn () => [[], []]),
+            ShortUrlsList::forTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listShortUrlsWithFilter($filter);
@@ -162,11 +164,49 @@ class ShlinkClientTest extends TestCase
     }
 
     /** @test */
+    public function listTagsWithFilterDelegatesCallToProperClient(): void
+    {
+        $filter = TagsFilter::create();
+        $listTags = $this->tagsClient->listTagsWithFilter($filter)->willReturn([]);
+
+        $this->shlinkClient->listTagsWithFilter($filter);
+
+        $listTags->shouldHaveBeenCalledOnce();
+    }
+
+    /** @test */
     public function listTagsWithStatsDelegatesCallToProperClient(): void
     {
-        $listTags = $this->tagsClient->listTagsWithStats()->willReturn([]);
+        $listTags = $this->tagsClient->listTagsWithStats()->willReturn(
+            TagsWithStatsList::forTupleLoader(static fn () => [[], []]),
+        );
 
         $this->shlinkClient->listTagsWithStats();
+
+        $listTags->shouldHaveBeenCalledOnce();
+    }
+
+    /** @test */
+    public function listTagsAndStatsDelegatesCallToProperClient(): void
+    {
+        $listTags = $this->tagsClient->listTagsAndStats()->willReturn(
+            TagsWithStatsList::forTupleLoader(static fn () => [[], []]),
+        );
+
+        $this->shlinkClient->listTagsAndStats();
+
+        $listTags->shouldHaveBeenCalledOnce();
+    }
+
+    /** @test */
+    public function listTagsWithStatsWithFilterDelegatesCallToProperClient(): void
+    {
+        $filter = TagsFilter::create();
+        $listTags = $this->tagsClient->listTagsWithStatsWithFilter($filter)->willReturn(
+            TagsWithStatsList::forTupleLoader(static fn () => [[], []]),
+        );
+
+        $this->shlinkClient->listTagsWithStatsWithFilter($filter);
 
         $listTags->shouldHaveBeenCalledOnce();
     }
@@ -207,7 +247,7 @@ class ShlinkClientTest extends TestCase
     {
         $identifier = ShortUrlIdentifier::fromShortCode('foo');
         $listVisits = $this->visitsClient->listShortUrlVisits($identifier)->willReturn(
-            VisitsList::forTupleLoader(fn () => [[], []]),
+            VisitsList::forTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listShortUrlVisits($identifier);
@@ -221,7 +261,7 @@ class ShlinkClientTest extends TestCase
         $identifier = ShortUrlIdentifier::fromShortCode('foo');
         $filter = VisitsFilter::create();
         $listVisits = $this->visitsClient->listShortUrlVisitsWithFilter($identifier, $filter)->willReturn(
-            VisitsList::forTupleLoader(fn () => [[], []]),
+            VisitsList::forTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listShortUrlVisitsWithFilter($identifier, $filter);
@@ -233,7 +273,7 @@ class ShlinkClientTest extends TestCase
     public function listTagVisitsDelegatesCallToProperClient(): void
     {
         $listVisits = $this->visitsClient->listTagVisits('foo')->willReturn(
-            VisitsList::forTupleLoader(fn () => [[], []]),
+            VisitsList::forTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listTagVisits('foo');
@@ -246,7 +286,7 @@ class ShlinkClientTest extends TestCase
     {
         $filter = VisitsFilter::create();
         $listVisits = $this->visitsClient->listTagVisitsWithFilter('foo', $filter)->willReturn(
-            VisitsList::forTupleLoader(fn () => [[], []]),
+            VisitsList::forTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listTagVisitsWithFilter('foo', $filter);
@@ -258,7 +298,7 @@ class ShlinkClientTest extends TestCase
     public function listOrphanVisitsDelegatesCallToProperClient(): void
     {
         $listVisits = $this->visitsClient->listOrphanVisits()->willReturn(
-            VisitsList::forOrphanVisitsTupleLoader(fn () => [[], []]),
+            VisitsList::forOrphanVisitsTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listOrphanVisits();
@@ -271,7 +311,7 @@ class ShlinkClientTest extends TestCase
     {
         $filter = VisitsFilter::create();
         $listVisits = $this->visitsClient->listOrphanVisitsWithFilter($filter)->willReturn(
-            VisitsList::forOrphanVisitsTupleLoader(fn () => [[], []]),
+            VisitsList::forOrphanVisitsTupleLoader(static fn () => [[], []]),
         );
 
         $this->shlinkClient->listOrphanVisitsWithFilter($filter);